refactor(core): 重构 FEValues 使用方式并优化线性求解器
- 在 Element 类中添加 create_fe_values 工厂方法，简化 FEValues 对象的创建- 更新 Current、Heat 和 Magnetic模块中的 assemble 函数，使用新的 FEValues 创建方式
-优化 LinearSolver 中的 BiCGSTAB 求解器，添加 IncompleteLUT 预条件器
- 改进 Exporter 中的 VTK 输出，支持多分量矢量数据和单元数据

+++ src/core/DOFManager.cpp
@@ -21,7 +21,7 @@ void DOFManager::build(const std::map<std::string, int>& field_orders) {
     // --- Pass 1: Assign DOFs to all vertex nodes ---
     for (const auto& node : mesh_.getNodes()) {
         for (size_t i = 0; i < variable_names_.size(); ++i) {
-            vertex_dof_map_[{node->getId(), i}] = equation_counter;
+            vertex_dof_map_[{node->getId(), static_cast<int>(i)}] = equation_counter;
             int num_components = variable_components_.at(variable_names_[i]);
             equation_counter += num_components;
         }
